Revert " create drop down, add arrow, decompose components (#42)"
This reverts commit 42b7d24914a2ae16ab75848bc87a4a7c1cca858c.
