+++ package.json
@@ -17,7 +17,6 @@
     "@fontsource/roboto": "^5.0.5",
     "@hookform/resolvers": "^3.1.1",
     "@radix-ui/react-checkbox": "^1.0.4",
-    "@radix-ui/react-dropdown-menu": "^2.0.5",
     "@radix-ui/react-label": "^2.0.2",
     "@radix-ui/react-radio-group": "^1.1.3",
     "@radix-ui/react-tabs": "^1.0.4",
