+++ src/components/auth/create-password-form/create-password-form.tsx
@@ -1,6 +1,6 @@
 import Button from '../../ui/button/button'
 import { Card } from '../../ui/card'
-import { Input } from '../../ui/drop-down-menu'
+import { Input } from '../../ui/Input'
 import { Typography } from '../../ui/Typography'
 
 import { CreatePasswordFormProps, useCreatePasswordForm } from './create-password-form'
