+++ src/components/auth/forgot-password-form/forgot-password-form.tsx
@@ -1,7 +1,7 @@
 import { PATH } from '../../../routes'
 import Button from '../../ui/button/button'
 import { Card } from '../../ui/card'
-import { Input } from '../../ui/drop-down-menu'
+import { Input } from '../../ui/Input'
 import { Typography } from '../../ui/Typography'
 
 import { ForgotPasswordFormProps, useForgotPasswordForm } from './forgot-password-form'
