+++ src/components/ui/controlled/controlled-input/controlled-input.tsx
@@ -1,6 +1,6 @@
 import { FieldValues, UseControllerProps, useController } from 'react-hook-form'
 
-import { Input, InputProps } from '../../drop-down-menu'
+import { Input, InputProps } from '../../Input'
 
 type ControlledInputProps<T extends FieldValues> = UseControllerProps<T> &
   Omit<InputProps, 'value' | 'onValueChange' | 'onChange' | 'id'>
