--- src/components/ui/drop-down-menu/index.ts
@@ -1,1 +1,0 @@
-export * from './drop-down-menu'
